+++ bin/typedef/linetv-station-data.js
@@ -1,0 +1,18 @@
+/**
+ * @typedef LINETvStationData
+ * @property {Array<LINETvClipData>} popularClip
+ * @property {Array<LINETvChannelData>} popularChannel
+ * @property {number} playCount
+ * @property {number} clipCount
+ * @property {string} stationId
+ * @property {number} likeitCount
+ * @property {number} channelCount
+ * @property {string} stationName
+ * @property {string} stationDesc
+ * @property {number} fanCount
+ * @property {string} stationLogo
+ * @property {string} stationBigBanner
+ * @property {{imgUrl: string; linkUrl: string}} stationBanner
+ */
+"use strict";
+//# sourceMappingURL=linetv-station-data.js.map